+++ ChangeLog.md
@@ -3,13 +3,14 @@
 LogicAppUnit Testing Framework:
 
 - Added new methods `TestRunner.WorkflowWasTerminated`, `TestRunner.WorkflowTerminationCode` and `TestRunner.WorkflowTerminationMessage` to allow the effects of a `Terminate` action in a workflow to be tested.
+- Mock responses can be configured using the fluent API in the test class initialization method, using the `WorkflowTestBase.AddMockResponse()` method. Mock responses configured using this method will be used by every test runner that is created in the same test class, and have a lower priority compared to the mock responses created using `ITestRunner.AddMockResponse()`.
 
 
 # 1.7.0 (27th July 2023)
 
 LogicAppUnit Testing Framework:
 
-- Mock responses can be configured using a fluent API, this includes the definition of the request matching conditions and the response. 
+- Mock responses can be configured using `ITestRunner.AddMockResponse()` and a fluent API, this includes the definition of the request matching conditions and the response. 
 - Removed public methods `ContentHelper.SerializeObject()`, `ContentHelper.DeserializeObject()` and `ContentHelper.JClone()`, these were for internal use only and are now obsolete.
 - Include the LogicAppUnit version at the end of the test log.
 - The maximum execution time for a workflow can now be set in the `testConfiguration.json` file using the `runner.maxWorkflowExecutionDuration` option. Previously this duration was hard-coded to 5 minutes. The default value for this option is 300 seconds (5 minutes).
@@ -35,8 +36,6 @@ LogicAppUnit Testing Framework:
 LogicAppUnit.Samples.LogicApps.Tests:
 
 - Added a `http-async-workflow` workflow and unit tests to demonstrate the use of the testing framework with HTTP triggers and asynchronous responses.
-
-Thanks to [@easchi](https://github.com/eashi) and [@atthevergeof](https://github.com/atthevergeof) for their contributions.
 
 
 # 1.5.0 (14th April 2023)